day_5/task_4/requirements.txt
streamlit
langchain
langchain-google-genai
chromadb
PyPDF2
tiktoken
python-dotenv
sentence-transformers
google-generativeai
